feat(ai): decompose AiPanel monolith into sub-components (v1-06)
- Extract AiPanel.tsx (2073→294 lines) into 9 focused sub-component files:
  AiMessageList, AiInputArea, AiPanelTabBar, AiEmptyState, AiUsageStats,
  AiProposalView, CodeBlock, aiPanelHelpers, useAiPanelEffects
- Add visual enhancements: accent border on AI messages (AC-8),
  mono font on code blocks (AC-9), severity-based ErrorGuideCard (AC-12),
  fade-in animation, spin-slow for Spinner
- Add AiPanelTabBar with Chat/History tabs and accent underline
- Add AiEmptyState with centered icon + prompt suggestions
- Add AiUsageStats with vertical token/cost display
- Guard tests: decomposition (≤300 lines, imports), visual tokens,
  focus-ring compliance, Lucide icon constraints
- i18n: tabChat, tabHistory, usageAnnotation keys (en + zh-CN)

Closes #1188
