+++ apps/desktop/renderer/src/features/ai/AiPanelTabBar.tsx
@@ -1,0 +1,50 @@
+import { useTranslation } from "react-i18next";
+
+type AiPanelTabBarProps = {
+  activeTab: "chat" | "history";
+  onTabChange: (tab: "chat" | "history") => void;
+};
+
+export function AiPanelTabBar(props: AiPanelTabBarProps): JSX.Element {
+  const { t } = useTranslation();
+  return (
+    <div
+      role="tablist"
+      className="flex border-b border-[var(--color-separator)] px-3"
+    >
+      <TabButton
+        label={t("ai.tabChat")}
+        active={props.activeTab === "chat"}
+        onClick={() => props.onTabChange("chat")}
+      />
+      <TabButton
+        label={t("ai.tabHistory")}
+        active={props.activeTab === "history"}
+        onClick={() => props.onTabChange("history")}
+      />
+    </div>
+  );
+}
+
+function TabButton(props: {
+  label: string;
+  active: boolean;
+  onClick: () => void;
+}): JSX.Element {
+  return (
+    // eslint-disable-next-line creonow/no-native-html-element -- specialized tab button
+    <button
+      type="button"
+      role="tab"
+      aria-selected={props.active}
+      onClick={props.onClick}
+      className={`px-3 py-2 text-[13px] font-medium transition-colors duration-[var(--duration-fast)] relative focus-ring ${
+        props.active
+          ? "text-[var(--color-fg-default)] border-b-2 border-[var(--color-accent)]"
+          : "text-[var(--color-fg-muted)] hover:text-[var(--color-fg-default)]"
+      }`}
+    >
+      {props.label}
+    </button>
+  );
+}
