+++ apps/desktop/renderer/src/i18n/locales/en.json
@@ -352,11 +352,14 @@
     "candidateSelected": "Selected",
     "regenerateAll": "Unsatisfied, regenerate all",
     "emptyHint": "Select text or enter a prompt to collaborate with AI",
+    "tabChat": "Chat",
+    "tabHistory": "History",
     "judgePass": "Quality check passed",
     "judgePartialSkipped": "Some checks were skipped",
     "usageOutput": "Output:",
     "usageSessionTotal": "Session total:",
     "usageCostEstimate": "Cost estimate:",
+    "usageAnnotation": "This conversation cost",
     "appliedSaved": "Applied & saved",
     "notConfigured": {
       "title": "Please configure AI service in Settings first",
