+++ apps/desktop/renderer/src/i18n/locales/zh-CN.json
@@ -352,11 +352,14 @@
     "candidateSelected": "已选择",
     "regenerateAll": "全部不满意，重新生成",
     "emptyHint": "选中文本或输入指令，开始与 AI 协作",
+    "tabChat": "对话",
+    "tabHistory": "历史",
     "judgePass": "质量校验通过",
     "judgePartialSkipped": "部分校验已跳过",
     "usageOutput": "输出：",
     "usageSessionTotal": "本会话累计：",
     "usageCostEstimate": "费用估算：",
+    "usageAnnotation": "本轮对话消耗",
     "appliedSaved": "已应用并保存",
     "notConfigured": {
       "title": "请先在设置中配置 AI 服务",
