+++ apps/desktop/renderer/src/styles/main.css
@@ -476,6 +476,35 @@ textarea:focus-visible {
   }
 }
 
+/* Fade in animation (AI empty state) */
+@keyframes fade-in {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@utility animate-fade-in {
+  animation: fade-in var(--duration-normal) var(--ease-default) forwards;
+  opacity: 0;
+}
+
+/* Slow spin animation (AI empty state sunburst) */
+@keyframes spin-slow {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
+}
+
+@utility animate-spin-slow {
+  animation: spin-slow 8s linear infinite;
+}
+
 /* Fade in up animation (Onboarding) */
 @keyframes fade-in-up {
   from {
